Main

Description

This application mimics a messaging app by sending 100 messages using 10 threads for 10 pre-defined users. The Application when run generates below mentioned outputs: - 100 instances of a Sender sending a message from the pre-defined pool to the Receiver. - Counts of the messages received by each specific user. - The Top 3 most active users, i.e., the Users who received the most messages -- {Additional Feature}

Written by Vishal

Documentation

main :: IO ()